rename package 'huffmancodin' to 'compression'
